Migrate shared preprod status check functionality to shared utils
