+++ src/sentry/preprod/api/models/project_preprod_build_details_models.py
@@ -16,7 +16,7 @@ from sentry.preprod.models import (
     PreprodComparisonApproval,
 )
 from sentry.preprod.snapshots.models import PreprodSnapshotComparison, PreprodSnapshotMetrics
-from sentry.preprod.vcs.status_checks.size.tasks import StatusCheckErrorType
+from sentry.preprod.vcs.status_checks.utils import StatusCheckErrorType
 
 logger = logging.getLogger(__name__)
